fix: errors on ProfileTest
